+++ consts/consts.js
@@ -6,6 +6,6 @@ export default {
     version: '0.1',
 
     fallback: 'http://helloworldx.ddns.net:8082',
-    fallbackChatServer: 'http://127.0.0.1:4006',
+    fallbackChatServer: 'http://helloworldx.ddns.net:9091',
 
 }
